requirements.txt
# Core dependencies
strands-agents
strands-agents-tools
mcp

# Utilities
python-dotenv

# MCP server runner
uvenv